+++ d2l-sequence-navigator/d2l-sequence-navigator.js
@@ -3,9 +3,9 @@ import '@polymer/polymer/polymer-legacy.js';
 import 'd2l-polymer-siren-behaviors/store/entity-behavior.js';
 import './d2l-outer-module.js';
 import 'd2l-accordion/d2l-accordion.js';
-import 'd2l-colors/d2l-colors.js';
+import '@brightspace-ui/core/components/colors/colors.js';
 import 'siren-entity/siren-entity.js';
-import '../../localize-behavior.js';
+import '../localize-behavior.js';
 import { html } from '@polymer/polymer/lib/utils/html-tag.js';
 import { mixinBehaviors } from '@polymer/polymer/lib/legacy/class.js';
 import { PolymerElement } from '@polymer/polymer/polymer-element.js';
