+++ demo/d2l-sequence-navigator/index.html
@@ -10,11 +10,11 @@
 	<script type="module">
 		import '@polymer/iron-demo-helpers/demo-pages-shared-styles';
 		import '@polymer/iron-demo-helpers/demo-snippet';
-		import '../../components/d2l-sequence-navigator/d2l-sequence-navigator.js';
+		import '../../d2l-sequence-navigator/d2l-sequence-navigator.js';
 		import 'd2l-typography/d2l-typography.js';
-		import '../../components/d2l-sequence-navigator/d2l-sequence-end.js';
-		import '../../components/d2l-sequence-navigator/d2l-lesson-header.js';
-		import '../../components/d2l-sequence-navigator/d2l-missed-activity.js';
+		import '../../d2l-sequence-navigator/d2l-sequence-end.js';
+		import '../../d2l-sequence-navigator/d2l-lesson-header.js';
+		import '../../d2l-sequence-navigator/d2l-missed-activity.js';
 		window.D2L.Siren.WhitelistBehavior._testMode(true);
 	</script>
