+++ demo/d2l-sequence-navigator/lesson-header.html
@@ -10,7 +10,7 @@
 	<script type="module">
 		import '@polymer/iron-demo-helpers/demo-pages-shared-styles';
 		import '@polymer/iron-demo-helpers/demo-snippet';
-		import '../../components/d2l-sequence-navigator/d2l-lesson-header.js';
+		import '../../d2l-sequence-navigator/d2l-lesson-header.js';
 	</script>
 
 	<custom-style>
