+++ demo/d2l-sequence-navigator/sequence.darkf.html
@@ -8,7 +8,7 @@
 		<script type="module">
 			import '@polymer/iron-demo-helpers/demo-pages-shared-styles';
 			import '@polymer/iron-demo-helpers/demo-snippet';
-			import '../../components/d2l-sequence-navigator/d2l-sequence-navigator.js';
+			import '../../d2l-sequence-navigator/d2l-sequence-navigator.js';
 			import 'd2l-typography/d2l-typography.js';
 		</script>
