+++ demo/d2l-sequences-content-html-multipage.html
@@ -1,14 +1,13 @@
 <html>
 <body>
-	<d2l-sequences-iterator id="prev" token="foo" icon="d2l-tier3:chevron-left-circle" previous=""></d2l-sequences-iterator>
-    <d2l-sequences-iterator id="next" token="foo" icon="d2l-tier3:chevron-right-circle" next=""></d2l-sequences-iterator>
+	<d2l-sequences-iterator id="prev" token="foo" icon="tier3:chevron-left-circle" previous=""></d2l-sequences-iterator>
+    <d2l-sequences-iterator id="next" token="foo" icon="tier3:chevron-right-circle" next=""></d2l-sequences-iterator>
 	<d2l-sequences-content-router href="../test/data/activity-file-html-multipage1.json" token="foo"></d2l-sequences-content-router>
 	<script type="module">
 		import { D2LSequencesContentFileHtml } from '../components/d2l-sequences-content-file-html.js'; // eslint-disable-line no-unused-vars
 		import { D2LSequencesIterator } from '../components/d2l-sequences-iterator.js'; // eslint-disable-line no-unused-vars
 		import SirenFixture from 'polymer-siren-test-helpers'; // eslint-disable-line no-unused-vars
 		import '../components/d2l-sequences-content-router.js';
-		import 'd2l-icons/tier3-icons.js';
 	</script>
 </body>
 </html>
