+++ demo/index.html
@@ -9,9 +9,7 @@
 			import '@polymer/iron-demo-helpers/demo-pages-shared-styles';
 			import '@polymer/iron-demo-helpers/demo-snippet';
 			import 'd2l-typography/d2l-typography.js';
-			import 'd2l-icons/d2l-icons.js';
-			import 'd2l-icons/tier2-icons.js';
-			import 'd2l-icons/tier3-icons.js';
+			import '@brightspace-ui/core/components/icons/icon.js';
 			import '../components/d2l-sequences-module-name.js';
 			import '../components/d2l-sequences-topic-name.js';
 			import '../components/d2l-sequences-iterator.js';
@@ -53,28 +51,28 @@
 			<h3>d2l-sequences-iterator (previous)</h3>
 			<demo-snippet>
 				<template>
-					<d2l-sequences-iterator previous icon="d2l-tier3:chevron-left-circle" href="data/l2-m1-activity1.json" token="foo"></d2l-sequences-iterator>
+					<d2l-sequences-iterator previous icon="tier3:chevron-left-circle" href="data/l2-m1-activity1.json" token="foo"></d2l-sequences-iterator>
 				</template>
 			</demo-snippet>
 
 			<h3>d2l-sequences-iterator (next)</h3>
 			<demo-snippet>
 				<template>
-					<d2l-sequences-iterator next icon="d2l-tier3:chevron-right-circle" href="data/l2-m1-activity1.json" token="foo"></d2l-sequences-iterator>
+					<d2l-sequences-iterator next icon="tier3:chevron-right-circle" href="data/l2-m1-activity1.json" token="foo"></d2l-sequences-iterator>
 				</template>
 			</demo-snippet>
 
 			<h3>d2l-sequences-iterator (previous disabled)</h3>
 			<demo-snippet>
 				<template>
-					<d2l-sequences-iterator previous icon="d2l-tier3:chevron-left-circle" href="" token="foo"></d2l-sequences-iterator>
+					<d2l-sequences-iterator previous icon="tier3:chevron-left-circle" href="" token="foo"></d2l-sequences-iterator>
 				</template>
 			</demo-snippet>
 
 			<h3>d2l-sequences-iterator (next disabled)</h3>
 			<demo-snippet>
 				<template>
-					<d2l-sequences-iterator next icon="d2l-tier3:chevron-right-circle" href="" token="foo"></d2l-sequences-iterator>
+					<d2l-sequences-iterator next icon="tier3:chevron-right-circle" href="" token="foo"></d2l-sequences-iterator>
 				</template>
 			</demo-snippet>
