+++ package.json
@@ -41,7 +41,7 @@
     "plop": "^2.0.0",
     "polymer-cli": "^1.9.10",
     "polymer-siren-test-helpers": "^2.0.0",
-    "sinon": "^5.0.7",
+    "sinon": "^7.5.0",
     "url-polyfill": "^1.1.3",
     "wct-browser-legacy": "^1.0.1",
     "whatwg-fetch": "^3.0.0"
@@ -67,13 +67,9 @@
     "@polymer/iron-collapse": "^3.0.0-pre.18",
     "@polymer/polymer": "^3.0.0",
     "d2l-accordion": "BrightspaceUI/accordion#semver:^1",
-    "d2l-button": "BrightspaceUI/button#semver:^5",
-    "d2l-colors": "BrightspaceUI/colors#semver:^4",
     "d2l-content-icons": "Brightspace/d2l-content-icons#semver:^3",
     "d2l-fetch": "Brightspace/d2l-fetch.git#semver:^2",
-    "d2l-icons": "BrightspaceUI/icons#semver:^6",
     "d2l-intl": "^2.1.0",
-    "d2l-link": "BrightspaceUI/link#semver:^5",
     "d2l-loading-spinner": "BrightspaceUI/loading-spinner#semver:^7",
     "d2l-localize-behavior": "BrightspaceUI/localize-behavior#semver:^2",
     "d2l-navigation": "BrightspaceUI/navigation#semver:^4",
