+++ test/d2l-sequence-navigator/d2l-activity-link.html
@@ -10,7 +10,7 @@
 		<script src="../../../wct-browser-legacy/browser.js"></script>
 		<script src="../../node_modules/chai-dom/chai-dom.js"></script>
 		<script type="module" src="../../../@polymer/polymer/polymer-legacy.js"></script>
-		<script type="module" src="../../components/d2l-sequence-navigator/d2l-activity-link.js"></script>
+		<script type="module" src="../../d2l-sequence-navigator/d2l-activity-link.js"></script>
 
 	</head>
 	<body>
@@ -52,7 +52,7 @@
 
 		<script type="module">
 import '@polymer/polymer/polymer-legacy.js';
-import '../../components/d2l-sequence-navigator/d2l-activity-link.js';
+import '../../d2l-sequence-navigator/d2l-activity-link.js';
 import SirenFixture from 'polymer-siren-test-helpers';
 describe('d2l-activity-link', () => {
