+++ test/d2l-sequence-navigator/d2l-completion-status.html
@@ -10,7 +10,7 @@
 		<script src="../../../wct-browser-legacy/browser.js"></script>
 		<script src="../../node_modules/chai-dom/chai-dom.js"></script>
 		<script type="module" src="../../../@polymer/polymer/polymer-legacy.js"></script>
-		<script type="module" src="../../components/d2l-sequence-navigator/d2l-completion-status.js"></script>
+		<script type="module" src="../../d2l-sequence-navigator/d2l-completion-status.js"></script>
 	</head>
 	<body>
 		<test-fixture id="CompletedActivityTestFixture">
@@ -27,7 +27,7 @@
 
 		<script type="module">
 import '@polymer/polymer/polymer-legacy.js';
-import '../../components/d2l-sequence-navigator/d2l-completion-status.js';
+import '../../d2l-sequence-navigator/d2l-completion-status.js';
 import SirenFixture from 'polymer-siren-test-helpers';
 
 describe('d2l-completion-status', () => {
