+++ test/d2l-sequence-navigator/d2l-inner-module.html
@@ -10,7 +10,7 @@
 		<script src="../../../wct-browser-legacy/browser.js"></script>
 		<script src="../../node_modules/chai-dom/chai-dom.js"></script>
 		<script type="module" src="../../../@polymer/polymer/polymer-legacy.js"></script>
-		<script type="module" src="../../components/d2l-sequence-navigator/d2l-inner-module.js"></script>
+		<script type="module" src="../../d2l-sequence-navigator/d2l-inner-module.js"></script>
 
 	</head>
 	<body>
@@ -38,7 +38,7 @@
 		</test-fixture>
 		<script type="module">
 import '@polymer/polymer/polymer-legacy.js';
-import '../../components/d2l-sequence-navigator/d2l-inner-module.js';
+import '../../d2l-sequence-navigator/d2l-inner-module.js';
 import SirenFixture from 'polymer-siren-test-helpers';
 
 describe('d2l-inner-module', () => {
