+++ test/d2l-sequence-navigator/d2l-lesson-header.html
@@ -10,7 +10,7 @@
 		<script src="../../../wct-browser-legacy/browser.js"></script>
 		<script src="../../node_modules/chai-dom/chai-dom.js"></script>
 		<script type="module" src="../../../@polymer/polymer/polymer-legacy.js"></script>
-		<script type="module" src="../../components/d2l-sequence-navigator/d2l-lesson-header.js"></script>
+		<script type="module" src="../../d2l-sequence-navigator/d2l-lesson-header.js"></script>
 		<style>
 			#DarkBackgroundFixture {
 				--d2l-asv-primary-color: #990000;
@@ -49,7 +49,7 @@
 		</test-fixture>
 		<script type="module">
 import '@polymer/polymer/polymer-legacy.js';
-import '../../components/d2l-sequence-navigator/d2l-lesson-header.js';
+import '../../d2l-sequence-navigator/d2l-lesson-header.js';
 import SirenFixture from 'polymer-siren-test-helpers';
 
 describe('d2l-lesson-header', () => {
