+++ test/d2l-sequence-navigator/d2l-sequence-navigator-activity.html
@@ -7,7 +7,7 @@
 	<script src="../../../@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
 	<script src="../../../wct-browser-legacy/browser.js"></script>
 	<script type="module" src="../../../@polymer/polymer/polymer-legacy.js"></script>
-	<script type="module" src="../../components/d2l-sequence-navigator/d2l-sequence-navigator.js"></script>
+	<script type="module" src="../../d2l-sequence-navigator/d2l-sequence-navigator.js"></script>
 </head>
 <body>
 
@@ -24,7 +24,7 @@
 </test-fixture>
 <script type="module">
 import '@polymer/polymer/polymer-legacy.js';
-import '../../components/d2l-sequence-navigator/d2l-sequence-navigator.js';
+import '../../d2l-sequence-navigator/d2l-sequence-navigator.js';
 import SirenFixture from 'polymer-siren-test-helpers';
 
 /* global SirenFixture */
